ProgrammingForAI-taba/education_analysis/requirements.txt
# Data processing and analysis
pandas>=1.3.0
numpy>=1.20.0
scipy==1.10.0

# Data visualization
matplotlib==3.7.0
seaborn==0.12.0
plotly==5.13.0

# Machine Learning
scikit-learn==1.2.0
nltk==3.8.1

# Database
psycopg2-binary>=2.9.1
pymongo>=4.0.1

# API and data collection
requests>=2.26.0
eurostat>=0.2.0
beautifulsoup4>=4.9.3
wbgapi>=1.0.0

# Environment and Jupyter
python-dotenv>=0.19.0
jupyter>=1.0.0
notebook>=6.4.0
pyarrow>=5.0.0  # 用于parquet文件支持

# Testing
pytest==7.3.1
